added another separate page for reset password
