+++ app/forgot-password.tsx
@@ -1,0 +1,77 @@
+import React, { useState } from "react";
+import {
+  View,
+  Text,
+  TextInput,
+  TouchableOpacity,
+  Alert,
+  Platform,
+  KeyboardAvoidingView,
+} from "react-native";
+import { useAuth } from "../contexts/AuthContext";
+import { useNavigation } from "@react-navigation/native";
+
+const ForgotPasswordScreen = () => {
+  const [email, setEmail] = useState("");
+  const [loading, setLoading] = useState(false);
+  const { resetPassword } = useAuth();
+  const navigation = useNavigation();
+
+  const handleResetPassword = async () => {
+    if (!email) {
+      Alert.alert("Error", "Please enter your email address");
+      return;
+    }
+
+    setLoading(true);
+    try {
+      const { error } = await resetPassword(email);
+      if (error) throw error;
+      Alert.alert("Success", "Check your email for the reset link!");
+      navigation.goBack(); 
+    } catch (error: any) {
+      Alert.alert("Error", error.message);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <KeyboardAvoidingView
+      className="flex-1 justify-center px-6 bg-gray-50"
+      behavior={Platform.OS === "ios" ? "padding" : "height"}
+    >
+      <View className="p-8 bg-white rounded-lg shadow-lg">
+        <Text className="mb-6 text-center text-2xl font-bold text-gray-800">
+          Reset Your Password
+        </Text>
+
+        <TextInput
+          className="px-4 py-3 text-gray-800 rounded-lg border border-gray-300 mb-4"
+          placeholder="Enter your email"
+          value={email}
+          onChangeText={setEmail}
+          keyboardType="email-address"
+          autoCapitalize="none"
+          autoCorrect={false}
+        />
+
+        <TouchableOpacity
+          className={`py-4 rounded-lg mb-4 ${loading ? "bg-gray-400" : "bg-blue-600"}`}
+          onPress={handleResetPassword}
+          disabled={loading}
+        >
+          <Text className="text-lg font-semibold text-center text-white">
+            {loading ? "Sending..." : "Send Reset Link"}
+          </Text>
+        </TouchableOpacity>
+
+        <TouchableOpacity onPress={() => navigation.goBack()}>
+          <Text className="text-center text-blue-600">Back to Sign In</Text>
+        </TouchableOpacity>
+      </View>
+    </KeyboardAvoidingView>
+  );
+};
+
+export default ForgotPasswordScreen;
